Configuration can now store long long coords

+++ src/Algorithm.h
@@ -10,10 +10,10 @@
 
 extern int MOWER_OVERLAP;
 extern int MAX_DEVIATION;
-extern int BASE_LON;
-extern int BASE_LAT;
-extern int BASE_EXIT_LON;
-extern int BASE_EXIT_LAT;
+extern long long BASE_LON;
+extern long long BASE_LAT;
+extern long long BASE_EXIT_LON;
+extern long long BASE_EXIT_LAT;
 
 //Gps
 extern long GpsGetLon();
